Change docker mode behaviour for opening ports to match the port on the host
Minifirewall was not perfect on docker hosts when one would expect to open ports to the outside world.
The iptables rules were written on the "destination" port leading to some oddities when when its different inside the container.

(ie: running a service exposed on port 7777 on the host, but listening inside the container with port 80.
For minifirewall, you had to open to the port 80. Leading to confusion (especially if you have more than
one container listening on the same port, but you don't want to expose them the same way))

This commit changes this behavior and introduce a new variable (DOCKER_LEGACY, "on" as default) to maintain the old behaviour.
This new behaviour (DOCKER='on' && 'DOCKER_LEGACY'='off') only works with the hosts ports when exposing sevices.

Switching the rules from :
${IPT} -I MINIFW-DOCKER-PUB -p tcp --dport "${dstport}" -j RETURN
to
${IPT} -I MINIFW-DOCKER-PUB -p tcp -m conntrack --ctorigdstport "${dstport}" -j RETURN

See #17